\name{as.data.frame.bayesian_first_aid}
\alias{as.data.frame.bayesian_first_aid}
\title{Returns the MCMC samples from a Bayesian First Aid fit as a data frame.}
\usage{
\method{as.data.frame}{bayesian_first_aid}(x, ...)
}
\arguments{
  \item{fit}{}
}
\value{
A data frame
}
\description{
Returns a dataframe with the MCMC samples from a Bayesian
First Aid Object. The samples does not include the burn in
phase and all chains are collapsed in the resulting data
frame.
}
\details{
Details details details
}

