\name{as.matrix.bayesian_first_aid}
\alias{as.matrix.bayesian_first_aid}
\title{Returns the MCMC samples from a Bayesian First Aid fit as a matrix.}
\usage{
\method{as.matrix}{bayesian_first_aid}(x, ...)
}
\arguments{
  \item{fit}{}
}
\value{
A matrix
}
\description{
Returns a matrix with the MCMC samples from a Bayesian
First Aid Object. The samples does not include the burn in
phase and all chains are collapsed in the resulting matrix.
}
\details{
Details details details
}

