\name{bayes.binom.test}
\alias{bayes.binom.test}
\title{Title title title}
\usage{
bayes.binom.test(x, n, comp.theta = 0.5, alternative = NULL,
  cred.mass = 0.95, p, conf.level)
}
\arguments{
  \item{x}{What is this param}
}
\value{
An object of type something...
}
\description{
Descritions description description
}
\details{
Details details details
}
\examples{
bayes.binom.test(5, 10)
}

