% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bayes.cor.test}
\alias{bayes.cor.test}
\alias{bayes.cor.test.default}
\alias{bayes.cor.test.formula}
\title{Bayesian First Aid Alternative to Pearson Correlation Test}
\usage{
bayes.cor.test(x, ...)

\method{bayes.cor.test}{default}(x, y, alternative = c("two.sided", "less",
  "greater"), method = c("pearson", "kendall", "spearman"), exact = NULL,
  cred.mass = 0.95, continuity = FALSE, n.iter = 15000,
  progress.bar = "text", conf.level, ...)

\method{bayes.cor.test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x,y}{numeric vectors of data values. \code{x} and \code{y} must have
the same length.}

\item{...}{not used}

\item{alternative}{ignored and is only retained in order to mantain
compatibility with \code{\link{cor.test}}}

\item{method}{ignored}

\item{exact}{ignored}

\item{cred.mass}{the amount of probability mass that will be contained in
reported credible intervals. This argument fills a similar role as
\code{conf.level} in \code{\link{cor.test}}.}

\item{continuity}{ignored}

\item{n.iter}{The number of iterations to run the MCMC sampling.}

\item{progress.bar}{The type of progress bar. Possible values are "text",
"gui", and "none".}

\item{formula}{a formula of the form \code{~ u + v}, where each of \code{u}
and \code{v} are numeric variables giving the data values for one sample.
The samples must be of the same length.}

\item{data}{an optional matrix or data frame containing the variables in the
formula \code{formula}. By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations to be
used.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs. Defaults to \code{getOption("na.action")}.}

\item{conf.level}{same as \code{cred.mass} and is only retained in order to
  mantain compatibility with \code{\link{cor.test}}.}
}
\value{
A list of class \code{bayes_cor_test} that contains information about
  the analysis. It can be further inspected using the functions
  \code{summary}, \code{plot}, \code{\link{diagnostics}} and
  \code{\link{model.code}}.
}
\description{
This is an implementation of an alternative to Person Correlation test as
implemented by \code{\link{cor.test}}. It accepts the same arguments as
\code{\link{cor.test}}. For more information regarding the model assumptions see:
\url{http://sumsar.net/blog/2014/03/bayesian-first-aid-pearson-correlation-test/}
}

