\name{bayes.poisson.test}
\alias{bayes.poisson.test}
\title{Bayesian First Aid Alternative to the Poisson Test}
\usage{
bayes.poisson.test(x, T = 1, r = 1, alternative = c("two.sided", "less",
  "greater"), conf.level = 0.95, n.iter = 15000)
}
\arguments{
  \item{x}{}

  \item{T}{}

  \item{r}{}

  \item{alternative}{}

  \item{conf.level}{}

  \item{n.iter}{}
}
\value{
A list of class \code{bayes_one_sample_poisson_test} or
\code{bayes_two_sample_poisson_test} that contains
information about the analysis. It can be further inspected
using the functions \code{summary}, \code{plot},
\code{\link{diagnostics}} and \code{\link{model.code}}.
}
\description{
Descritions description description
}
\details{
Details details details
}

