\name{jags_one_sample_poisson_test}
\alias{jags_one_sample_poisson_test}
\title{Generate MCMC samples from the Bayesian First Aid one sample poisson model using JAGS.}
\usage{
  jags_one_sample_poisson_test(x, t)
}
\arguments{
  \item{x}{the number of events}

  \item{t}{The number of time steps during the \code{x}
  events were observed.

  #'}
}
\value{
  An object of type \code{mcmc.list} (defined by the
  \code{coda} package) that contains the MCMC samples from
  the model.
}
\description{
  Descritions description description
}
\details{
  Details details details
}

