\name{jags_two_sample_t_test}
\alias{jags_two_sample_t_test}
\title{Adapted from John Kruschke's original BEST code.}
\usage{
  jags_two_sample_t_test(x, y, n.adapt = 1000,
    n.chains = 3, n.update = 1000, n.iter = 1000, thin = 1)
}
\description{
  Adapted from John Kruschke's original BEST code.
}

