\name{model.code}
\alias{model.code}
\title{Prints code that replicates the model you just ran.}
\usage{
model.code(fit)
}
\arguments{
  \item{x}{What is this param}
}
\value{
An object of type something...
}
\description{
This is good if you better want to understand how the model
is implemented or if you want to run a modified version of
the code.
}
\details{
Details details details
}

