\name{model_diagram}
\alias{model_diagram}
\title{Plots the model diagram in the style of Kruschke(2010)}
\usage{
  model_diagram(fit)
}
\description{
  Plots the model diagram in the style of Kruschke(2010)
}
\references{
  Kruschke, J. (2010) Doing Bayesian Data Analysis: A
  Tutorial Introduction with R. Academic Press.
}

