% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_cor_test.R
\name{plot.bayes_cor_test}
\alias{plot.bayes_cor_test}
\title{Summary plot of a bayes.cor.test object}
\usage{
\method{plot}{bayes_cor_test}(x, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{the resulting object from a \code{\link{bayes.cor.test}} run}

\item{xlim, ylim}{the limits of the scatterplot, if NULL then reasonable limits will be calculated automatically.}
}
\description{
Produces a histogram of the posterior of the correlation parameter rho, and a scatterplot with the data and the posterior predictive density.
}

