\name{plotPost}
\alias{plotPost}
\title{Author John Kruschke}
\usage{
  plotPost(param_sample_vec, cred_mass = 0.95,
    comp_val = NULL, HDI_text_place = 0.7, ROPE = NULL,
    yaxt = NULL, ylab = NULL, xlab = NULL, cex.lab = NULL,
    cex = NULL, xlim = NULL, main = NULL, col = NULL,
    border = NULL, show_mode = F, show_curve = F,
    breaks = NULL, ...)
}
\description{
  Author John Kruschke
}

