% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.matrix.bayesian_first_aid}
\alias{as.matrix.bayesian_first_aid}
\title{Returns the MCMC samples from a Bayesian First Aid fit as a matrix.}
\usage{
\method{as.matrix}{bayesian_first_aid}(x, ...)
}
\arguments{
\item{x}{The output from a Bayesian First Aid model.}

\item{...}{Not used}
}
\value{
A matrix with one column per parameter.
}
\description{
Returns a matrix with the MCMC samples from a Bayesian First Aid Object. The
samples does not include the burn in phase and all chains are collapsed in
the resulting matrix.
}
\details{
This is just a wrapper for the \code{as.matrix} from the coda package.
Therefor, paramameters with the same name but different index will be column
names that includes brackets (for example, "mu[1]", "mu[2]", "sigma[1,2],
etc.)
}
\examples{
fit <- bayes.t.test(rnorm(10), rnorm(10))
d <- as.matrix(fit)
mean(d[,"mu_x"])
}

