\name{bayes.cor.test}
\alias{bayes.cor.test}
\alias{bayes.cor.test.default}
\alias{bayes.cor.test.formula}
\title{Title title title}
\usage{
bayes.cor.test(x, ...)

\method{bayes.cor.test}{default}(x, y, alternative = c("two.sided", "less",
  "greater"), method = c("pearson", "kendall", "spearman"), exact = NULL,
  conf.level = 0.95, continuity = FALSE, n.iter = 15000, ...)

\method{bayes.cor.test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
  \item{x}{What is this param}
}
\value{
An object of type something...
}
\description{
Descritions description description
}
\details{
Details details details
}

