% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bayes.cor.test}
\alias{bayes.cor.test}
\alias{bayes.cor.test.default}
\alias{bayes.cor.test.formula}
\title{Bayesian First Aid Alternative to Pearson Correlation Test}
\usage{
bayes.cor.test(x, ...)

\method{bayes.cor.test}{default}(x, y, alternative = c("two.sided", "less",
  "greater"), method = c("pearson", "kendall", "spearman"), exact = NULL,
  cred.mass = 0.95, continuity = FALSE, n.iter = 15000,
  progress.bar = "text", conf.level, ...)

\method{bayes.cor.test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x,y}{numeric vectors of data values. \code{x} and \code{y} must have
the same length.}

\item{...}{not used}

\item{alternative}{ignored and is only retained in order to mantain
compatibility with \code{\link{cor.test}}}

\item{method}{ignored}

\item{exact}{ignored}

\item{cred.mass}{the amount of probability mass that will be contained in
reported credible intervals. This argument fills a similar role as
\code{conf.level} in \code{\link{cor.test}}.}

\item{continuity}{ignored}

\item{n.iter}{The number of iterations to run the MCMC sampling.}

\item{progress.bar}{The type of progress bar. Possible values are "text",
"gui", and "none".}

\item{formula}{a formula of the form \code{~ u + v}, where each of \code{u}
and \code{v} are numeric variables giving the data values for one sample.
The samples must be of the same length.}

\item{data}{an optional matrix or data frame containing the variables in the
formula \code{formula}. By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations to be
used.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs. Defaults to \code{getOption("na.action")}.}

\item{conf.level}{same as \code{cred.mass} and is only retained in order to
  mantain compatibility with \code{\link{cor.test}}.}
}
\value{
A list of class \code{bayes_cor_test} that contains information about
  the analysis. It can be further inspected using the functions
  \code{summary}, \code{plot}, \code{\link{diagnostics}} and
  \code{\link{model.code}}.
}
\description{
This is an implementation of an alternative to Person Correlation test as
implemented by \code{\link{cor.test}}. It accepts the same arguments as
\code{\link{cor.test}}. For more information regarding the model assumptions see:
\url{http://sumsar.net/blog/2014/03/bayesian-first-aid-pearson-correlation-test/}
}
\examples{
# Data from Hollander & Wolfe (1973), p. 187f.
# This example is borrowed from the documentation for cor.test().

# Assessment of tuna quality.  We compare the Hunter L measure of
#  lightness to the averages of consumer panel scores (recoded as
#  integer values from 1 to 6 and averaged over 80 such values) in
#  9 lots of canned tuna.
x <- c(44.4, 45.9, 41.9, 53.3, 44.7, 44.1, 50.7, 45.2, 60.1)
y <- c( 2.6, 3.1, 2.5, 5.0, 3.6, 4.0, 5.2, 2.8, 3.8)

# First a classical correlation test:
cor.test(x, y)

# And here is the Bayesian first aid alternative:
bayes.cor.test(x, y)

# Save the output into a variable for easy plotting and further inspection:
fit <- bayes.cor.test(x, y)
plot(fit)
summary(fit)
model.code(fit)
}

