% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bayes.cor.test}
\alias{bayes.cor.test}
\alias{bayes.cor.test.default}
\alias{bayes.cor.test.formula}
\title{Title title title}
\usage{
bayes.cor.test(x, ...)

\method{bayes.cor.test}{default}(x, y, alternative = c("two.sided", "less",
  "greater"), method = c("pearson", "kendall", "spearman"), exact = NULL,
  cred.mass = 0.95, continuity = FALSE, n.iter = 15000,
  progress.bar = "text", conf.level, ...)

\method{bayes.cor.test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{}

\item{...}{}

\item{y}{}

\item{alternative}{}

\item{method}{}

\item{exact}{}

\item{cred.mass}{}

\item{continuity}{}

\item{n.iter}{}

\item{progress.bar}{The type of progress bar. Possible values are "text",
"gui", and "none".}

\item{formula}{}

\item{data}{}

\item{subset}{}

\item{na.action}{}

\item{conf.level}{}
}
\value{
A list of class \code{bayes_cor_test} that contains information about
  the analysis. It can be further inspected using the functions
  \code{summary}, \code{plot}, \code{\link{diagnostics}} and
  \code{\link{model.code}}.
}
\description{
Descritions description description
}
\details{
Details details details
}

