\name{bayes.cor.test}
\alias{bayes.cor.test}
\alias{bayes.cor.test.default}
\alias{bayes.cor.test.formula}
\title{Title title title}
\usage{
bayes.cor.test(x, ...)

\method{bayes.cor.test}{default}(x, y, alternative = c("two.sided", "less",
  "greater"), method = c("pearson", "kendall", "spearman"), exact = NULL,
  conf.level = 0.95, continuity = FALSE, n.iter = 15000, ...)

\method{bayes.cor.test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
  \item{x}{}

  \item{...}{}

  \item{y}{}

  \item{alternative}{}

  \item{method}{}

  \item{exact}{}

  \item{conf.level}{}

  \item{continuity}{}

  \item{n.iter}{}

  \item{formula}{}

  \item{data}{}

  \item{subset}{}

  \item{na.action}{}
}
\value{
A list of class \code{bayes_cor_test} that contains
information about the analysis. It can be further inspected
using the functions \code{summary}, \code{plot},
\code{\link{diagnostics}} and \code{\link{model.code}}.
}
\description{
Descritions description description
}
\details{
Details details details
}

