\name{bayes.poisson.test}
\alias{bayes.poisson.test}
\title{Title title title}
\usage{
bayes.poisson.test(x, T = 1, r = 1, alternative = c("two.sided", "less",
  "greater"), conf.level = 0.95, n.iter = 15000)
}
\arguments{
  \item{x}{}

  \item{T}{}

  \item{r}{}

  \item{alternative}{}

  \item{conf.level}{}

  \item{n.iter}{}
}
\value{
An object of type something...
}
\description{
Descritions description description
}
\details{
Details details details
}

