% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diagnostics}
\alias{diagnostics}
\title{Plots and prints diagnostics regarding the convergence of the model.}
\usage{
diagnostics(fit)
}
\arguments{
\item{fit}{The output from a Bayesian First Aid model.}
}
\description{
Plots and prints diagnostics regarding the convergence of the model.
}

