\name{diagnostics}
\alias{diagnostics}
\title{Plots and prints diagnostics regarding the convergence of the model.}
\usage{
diagnostics(fit)
}
\arguments{
  \item{fit}{}
}
\value{
An object of type something...
}
\description{
Descritions description description
}
\details{
Details details details
}

