% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{diagnostics}
\alias{diagnostics}
\title{Plots and prints diagnostics regarding the convergence of the model.}
\usage{
diagnostics(fit)
}
\arguments{
\item{fit}{The output from a Bayesian First Aid model.}
}
\description{
Plots and prints diagnostics regarding the convergence of the model.
}

