% Generated by roxygen2 (4.0.1): do not edit by hand
\name{model.code}
\alias{model.code}
\title{Prints code that replicates the model you just ran.}
\usage{
model.code(fit)
}
\arguments{
\item{fit}{}
}
\value{
An object of type something...
}
\description{
This is good if you better want to understand how the model is
implemented or if you want to run a modified version of the code.
}
\details{
Details details details
}

