% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{model.code}
\alias{model.code}
\title{Prints code that replicates the model you just ran.}
\usage{
model.code(fit)
}
\arguments{
\item{fit}{The output from a Bayesian First Aid model.}
}
\description{
This is good if you better want to understand how the model is
implemented or if you want to run a modified version of the code.
}

